move functions into lib
